Changed Int to Float

+++ can.js
@@ -1,7 +1,7 @@
 function mult() {
             var turkish = document.getElementById('try').value;
             var bitcoin = document.getElementById('btc').value;
-            var result = parseInt(turkish) / parseInt(bitcoin);
+            var result = parseFloat(turkish) / parseFloat(bitcoin);
             if (!isNaN(result)) {
                 document.getElementById('rate').value = result;
             }
